+++ rustls-wolfcrypt-provider/src/lib.rs
@@ -27,8 +27,7 @@ pub mod aead {
 pub mod sign {
     pub mod ecdsa;
     pub mod eddsa;
-    pub mod rsapkcs1;
-    pub mod rsapss;
+    pub mod rsa;
 }
 use crate::aead::{aes128gcm, aes256gcm, chacha20};
 
@@ -94,10 +93,7 @@ impl rustls::crypto::KeyProvider for Provider {
         // Define supported algorithms as closures
         let algorithms: SigningAlgorithms = vec![
             Box::new(|key| sign::ecdsa::EcdsaSigningKey::try_from(key).map(|x| Arc::new(x) as _)),
-            Box::new(|key| sign::rsapss::RsaPssPrivateKey::try_from(key).map(|x| Arc::new(x) as _)),
-            Box::new(|key| {
-                sign::rsapkcs1::RsaPkcs1PrivateKey::try_from(key).map(|x| Arc::new(x) as _)
-            }),
+            Box::new(|key| sign::rsa::RsaPrivateKey::try_from(key).map(|x| Arc::new(x) as _)),
             Box::new(|key| sign::eddsa::Ed25519PrivateKey::try_from(key).map(|x| Arc::new(x) as _)),
         ];
